requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-dotenv
jinja2
faiss-cpu
sentence-transformers
langchain
jinja2